upload file fixed

+++ src/main/java/tn/esprit/services/RefreshTokenService.java
@@ -25,7 +25,7 @@ public class RefreshTokenService {
 	  }
 	  public RefreshToken createRefreshToken(int userId) {
 	    RefreshToken refreshToken = new RefreshToken();
-	    refreshToken.setUser(userRepository.findById(userId).get());
+	    refreshToken.setUser(userRepository.findById((int) userId).get());
 	    refreshToken.setExpiryDate(Instant.now().plusMillis(refreshTokenDurationMs));
 	    refreshToken.setToken(UUID.randomUUID().toString());
 	    refreshToken = refreshTokenRepository.save(refreshToken);
